requirements.txt
Django==4.1.5
psycopg2==2.9.5
numpy==1.22.4
pandas==1.5.3
kaggle==1.5.13
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-apscheduler==0.6.2
APScheduler~=3.10.1
django-environ==0.10.0
djangorestframework-simplejwt==5.2.2
catboost==1.1.1
scikit-learn==1.2.2
joblib==1.2.0